receive: convert file path to local form when registering.
This matches the same transformation as is done in `DiskWriter.Handle`, which
eventually gets passed to the `AsyncDataCb` AKA `receiver.asyncDataFunc` which
performs the lookup.

This matters on Windows where `filepath.FromSlash` will convert `/` into `\`.
I've tested this (in the context of buildkit) with a Linux `sender` and Windows
`receiver`, I've no capability to test the opposite direction.

Signed-off-by: Ian Campbell <ijc@docker.com>