diskwriter: pass wire/unix paths to callbacks.
The callbacks expect to operate on the wire form which is standardized on Unix
format.

For example `receiver.asyncDataFunc`, which is used as `AsyncDataCb` expects to
use the path as a key on a map lookup for an insertion done in the
`PACKET_STAT` case of `receiver.run` using the wire form.

This matters on Windows where `filepath.FromSlash` will convert `/` into `\`.

Only the `destPath` needs to be converted to the local form. I have confirmed
that all subsequent uses of `p` in `DiskWriter.Handle` are passing it to one of
the callbacks (either directly or via `DiskWriter.processChange` or
`DiskWriter.requestAsyncFileData`, that latter just calls `processChange`) as a
key or for error messages.

The callbacks reachable in this way are `NotifyCallback`, `SyncDataCb` and
`AsyncDataCb`, which are set to `recevier.notifyHashed`, `newWriteToFunc`
(ignoring `nopWriteTo`) functions and `receiver.asyncDataFunc` respectively.

The `newWriteToFunc` is adjusted to cope with this change and
`receive.asyncDataFunc` already expected a canonical form for path and does thr
correct thing.

The `receiver.notifyHashed` callback is user-supplied via
`RecieveOpt.NotifyHashed`. I have looked at the user in
`github.com/moby/buildkit/session/filesync/diffcopy.go` which uses the
`HandleChange` on a supplied `CacheUpdater`. One candidate for a supplied
object meeting this interface is
`github.com/moby/buildkit/cache/contenthash/cacheContext`. The `HandleChange`
callback there does an immediate `filepath.ToSlash` which I think will need to
be removed when buildkit is updated, I've been unable to determine if the
consequences of not doing so are more serious than an artificial cache-bust.

I've tested this (in the context of buildkit) with a Linux `sender` and Windows
`receiver`, I've no capability to test the opposite direction.

Signed-off-by: Ian Campbell <ijc@docker.com>