Add IncludePatterns and ExcludePatterns options for Copy
Allow include patterns and exclude patterns to be specified, similarly
to Walker.

There is a bit of extra complexity to handle the case of a pattern like
`a/*/c`. In this case, creating the directories `a` and `a/b` may need
to be delayed until we encounter `a/b/c`.