Use PatternMatcher for copy include patterns as well as exclude patterns
This changes the logic for copy's "include patterns" to use
PatternMatcher, so it behaves the same as "exclude patterns". This is a
net simplification and provides more consistency between the two
parameters (for example, include patterns can now specify negative
matches), but is a bit worse performance-wise (we can't skip a directory
which fails to partially match a pattern, because files inside it may
match the pattern).

Signed-off-by: Aaron Lehmann <alehmann@netflix.com>