hack/dockerfiles: use GO_VERSION for linter and go mod validation
This allows the version to be set through a build-arg, and makes sure that
the linter lints on the expected go version.

Note that go mod results may differ between go versions. If this becomes a
problem, we should use a fixed version for that validation (added a comment
in the bakefile).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>