Handle EPERM errors when walking excluded dirs
A recent change to `Walk` causes it to walk dirs that matched an exclude
pattern, in cases where there are exceptions to the exclude pattern (do
we have better terminology for this?). This was done in order to support
exclusions like:

    foo
    !**/bar

In that example, we want to walk `foo` even though it's excluded,
because there might be a file or directory `bar` somewhere underneath
it.

However, this change in behavior surfaced "permission denied" errors
that didn't show up before. In the case that there is no permission to
list contents of "foo", that now results in an error.

To handle this properly, we need to ignore EPERM errors on dirs that we
are only accessing to search for these hypothetical nested files/dirs
which might end up included.

A followup change will optimize this behavior by avoiding the search in
cases where such exclusions can't possibly apply based on the patterns
involved.

Signed-off-by: Aaron Lehmann <alehmann@netflix.com>