copy: don't copy file info of non-existent dir.
Before this, if a directory was excluded, copy would still end up
getting called with overwriteTargetMetadata=true in the code path where
subdirs of the root of the src are iterated over. This meant that even
though the target directory was never created and never existed, the
copyFileInfo bool would stay its default value of true and the code
would try to copy file info to a directory that doesn't exist.

This update just has copyFileInfo default to whether the target should
actually be included.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>