send: always use buf in CopyBuffer
CopyBuffer may not always use the provided buffer, in the case that the
src implements WriterTo. So, in the case that the io.ReadCloser returned
by fs.Open implements this method, the packet size limit will not be
enforced, creating an oversized packet that results in an error on the
receiver.

To avoid this, we can create an anonymous structure to wrap the src,
which prevents the shortcut taken by the standard library.

Signed-off-by: Justin Chadwell <me@jedevc.com>