Implement proper renameFile on Windows
Symlinks on Windows are implemented using NTFS reparse points. Reparse
points are also used when creating volume mounts on Windows. As far as
golang is concerned, mountpoints and symlinks look the same. This change
makes the distinction between the two and allows safely renaming
symlinks, while erring out for anything else that looks like a symlink.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>