Use RunWithPrivileges
The EnableProcessPrivileges function will enable a certain set of
privileges for the entire process. I am unsure if enabling/disabling
privileges is tracked with some sort of reference counting, but if it's
not, when we release those privileges, we disable them for the entire
process. This means that if some other go routine requires them, we may
end up with undesirable results.

The RunWithPrivileges locks the thread in which the function runs, and
enables those privileges only in that thread.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>