filter: allow SkipDir return with lazy parents
When using include/exclude options in a file walk, we lazily store the
parents, and only later call the target function with them when we find
a file inside. This handles the case where `**/*.go` should only walk
directories that contain `.go` files, and not every directory in the
tree.

However, when storing lazy parents, we weren't appropriately handling
the `SkipDir` value that could be returned. We were making two incorrect
assumptions:

1. An error value from `fn()` was always an error case - this isn't
   true, since we have control error values like `SkipDir`. Because of
   this, we need to ensure that `calledFn` is *always* set to true when
   calling `fn` (even in an error case).
2. We can return `SkipDir` to skip this parent directory. This isn't
   *quite* right, since returning `SkipDir` would return it only for the
   child directory - we need to make sure it's called everytime for each
   child directory of this parent directory, by storing it as a boolean
   `skipFn` field.

With these changes, the new test passes, and we don't end up with
duplicate parent calls, as was happening before:

	Error:          Not equal:
			expected: []string{"foo"}
			actual  : []string{"foo", "foo"}

Signed-off-by: Justin Chadwell <me@jedevc.com>