Avoid using syscall for errors on unix
syscall is deprecated and non updated library whcih should be replaced
by new one with name x/sys.

Anyway, it may work on some platform and some errors migth be missed on
some. A good example is OpenBSD which has lack of syscall.EBADMSG on
amd64 and 386 but has on arm64.

Here I moved all errors to two files: one for unix which uses sys.unix,
and one for everything else which uses syscall as fallback.