fix: clear ModeIrregular on directories on Windows
Since go1.23, directories with an unknown reparse tag will have the
irregular bit set. We do not know exactly which scenario causes this, but
there are many different reparse tags that can be set and go only handles
3 of them.

This causes issues with other parts of the code because in Unix a directory
cannot have this bit set.

Add a workaround to clear this bit for directories on Windows.